Add files via upload
Creating a query to count how many users are added daily, 
accounting for user accounts that have been merged and/or deleted. 


Created a subtable that counted the total number of users added each day. 
Created a separate subtable to count the total number of deleted users and one for merged users. Then filtered them by day.
Joined the deleted users' table and the merged user's table to the original subtable in order to create new table.
Coalesced columns to include 0's instead of empty records. 
Added net users column by taking the difference of the new users and the merged and deleted users.